// src/types.ts
export type Theme = {
  id: string;
  name: string;
  description: string;
  classes: {
    app: string;
    sidebar: string;
    sidebarItemActive: string;
    sidebarItemHover: string;
    header: string;
    navLink: string;
    hero: string;
    heroText: string;
    section: string;
    sectionAlt: string;
    heading: string;
    card: string;
    button: string;
  };
};